pyproject.toml
[project]
name = "blog"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-frontmatter>=1.1.0",
    "ipython>=9.4.0",
    "markdown>=3.8.2",
]
